+++ Content.Server/_ES/Stagehand/Components/ESUsernameEntityNameComponent.cs
@@ -1,10 +1,8 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._ES.Stagehand.Components;
+namespace Content.Server._ES.Stagehand.Components;
 
 /// <summary>
 /// Used for entities which show the player's username when they are occupied, rather than an IC name.
 /// </summary>
-[RegisterComponent, NetworkedComponent]
+[RegisterComponent]
 [Access(typeof(ESUsernameEntityNameSystem))]
 public sealed partial class ESUsernameEntityNameComponent : Component;
