+++ Content.Shared/_ES/Stagehand/Components/ESUsernameEntityVisuals.cs
@@ -1,0 +1,9 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Shared._ES.Stagehand.Components;
+
+[Serializable, NetSerializable]
+public enum ESUsernameEntityVisuals : byte
+{
+    Admin,
+}
