--- Content.Shared/_ES/Stagehand/ESUsernameEntityNameSystem.cs
@@ -1,19 +1,0 @@
-using Content.Shared._ES.Stagehand.Components;
-using Robust.Shared.Player;
-
-namespace Content.Shared._ES.Stagehand;
-
-public sealed partial class ESUsernameEntityNameSystem : EntitySystem
-{
-    [Dependency] private MetaDataSystem _metaData = default!;
-
-    /// <inheritdoc/>
-    public override void Initialize()
-    {
-        SubscribeLocalEvent<ESUsernameEntityNameComponent, PlayerAttachedEvent>(OnPlayerAttached);
-    }
-    private void OnPlayerAttached(Entity<ESUsernameEntityNameComponent> ent, ref PlayerAttachedEvent args)
-    {
-        _metaData.SetEntityName(ent, args.Player.Name);
-    }
-}
