+++ Resources/Textures/_ES/Mobs/Ghosts/theatergoer.rsi/meta.json
@@ -14,6 +14,10 @@
       "directions": 4
     },
     {
+      "name": "admin",
+      "directions": 4
+    },
+    {
       "name": "stagehand",
       "directions": 4
     },
